refactor: rename files and minor clean up
