+++ cpp/serialize/utils.h
